feat(desktop): shrink updates row to settings-inline pill
Replace the full-width UpdatesRow with UpdatesPill, a compact mono
pill in the sidebar's bottom settings cluster: progress ring +
percent while downloading, pulsing green '↑ update' when ready
(version in tooltip), micro ASCII loader while installing, red
'↻ retry' on failure. Mounted in both DashboardSidebar (v2) and the
v1 WorkspaceSidebar footer; icon-only variant when collapsed.
