+++ apps/desktop/src/renderer/components/UpdatesPill/index.ts
@@ -1,0 +1,1 @@
+export { UpdatesPill } from "./UpdatesPill";
