+++ apps/desktop/src/renderer/components/UpdatesPill/useAutoUpdateStatus.ts
