--- apps/desktop/src/renderer/components/UpdatesRow/index.ts
@@ -1,1 +1,0 @@
-export { UpdatesRow } from "./UpdatesRow";
