+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/DashboardSidebar.tsx
@@ -24,7 +24,7 @@ import { useMatchRoute, useNavigate } from "@tanstack/react-router";
 import { memo, useCallback, useEffect, useMemo, useState } from "react";
 import { createPortal } from "react-dom";
 import { HiOutlineCog6Tooth } from "react-icons/hi2";
-import { UpdatesRow } from "renderer/components/UpdatesRow";
+import { UpdatesPill } from "renderer/components/UpdatesPill";
 import { useHotkeyDisplay } from "renderer/hotkeys";
 import { useDashboardSidebarState } from "renderer/routes/_authenticated/hooks/useDashboardSidebarState";
 import { useLocalHostService } from "renderer/routes/_authenticated/providers/LocalHostServiceProvider";
@@ -244,7 +244,6 @@ export function DashboardSidebar({
 									projectName={activeV2Project.name}
 								/>
 							)}
-							<UpdatesRow isCollapsed={isCollapsed} />
 							<div
 								className={cn(
 									"border-t border-border",
@@ -298,6 +297,7 @@ export function DashboardSidebar({
 									</button>
 								)}
 
+								<UpdatesPill isCollapsed={isCollapsed} />
 								<DashboardSidebarHelpMenu isCollapsed={isCollapsed} />
 							</div>
 						</div>
